+++ example/a.ts
@@ -1,17 +1,14 @@
 import { Elysia, t } from '../src'
-import { req } from '../test/utils'
 
-const asyncPlugin = Promise.resolve(new Elysia({ name: 'AsyncPlugin' }))
+const app = new Elysia().mount('/v1/*', (request) => {
+	return Response.json({
+		method: request.method,
+		path: request.url
+	})
+})
 
-const plugin = new Elysia({ name: 'Plugin' })
-	.use(asyncPlugin)
-	.get('/plugin', () => 'GET /plugin')
+const response = await app
+	.handle(new Request('http://localhost/v1/hello'))
+	.then((x) => x.json())
 
-const app = new Elysia({ name: 'App' })
-	.use(plugin)
-	.get('/foo', () => 'GET /foo')
-	.listen(3000)
-
-const response = await app.handle(new Request('http://localhost/plugin'))
-const text = await response.text()
-console.log(text) // 'GET /plugin'
+// console.log(app.routes[0].composed.toString())